requirements.txt
fastapi
uvicorn
pydantic
requests
chainlit
pytest
azure-identity
pydantic-settings
azure-monitor-query 